+++ examples/cryptocurrency-advanced/backend/src/api.rs
@@ -15,16 +15,15 @@
 //! Cryptocurrency API.
 
 use bodyparser;
+use exonum::{api::{Api, ApiError},
+             blockchain::{self, BlockProof, Blockchain, Transaction, TransactionSet},
+             crypto::{Hash, PublicKey},
+             helpers::Height,
+             node::TransactionSend,
+             storage::{ListProof, MapProof}};
 use iron::prelude::*;
 use router::Router;
 use serde_json;
-
-use exonum::api::{Api, ApiError};
-use exonum::blockchain::{self, BlockProof, Blockchain, Transaction, TransactionSet};
-use exonum::crypto::{Hash, PublicKey};
-use exonum::helpers::Height;
-use exonum::node::TransactionSend;
-use exonum::storage::{ListProof, MapProof};
 
 use std::fmt;