+++ exonum/src/encoding/mod.rs
@@ -105,17 +105,15 @@
 //!
 //! [`field_size()`]: ./trait.Field.html#tymethod.field_size
 
-pub use self::error::Error;
-pub use self::fields::Field;
 #[cfg(feature = "float_serialize")]
 pub use self::float::{F32, F64};
-pub use self::segments::SegmentField;
+pub use self::{error::Error, fields::Field, segments::SegmentField};
 
 #[macro_use]
 pub mod serialize;
 
-use std::convert::From;
-use std::ops::{Add, Div, Mul, Sub};
+use std::{convert::From,
+          ops::{Add, Div, Mul, Sub}};
 
 mod error;
 #[macro_use]