+++ exonum/src/helpers/fabric/shared.rs
@@ -16,11 +16,9 @@
 
 use toml;
 
-use std::collections::BTreeMap;
-use std::net::SocketAddr;
+use std::{collections::BTreeMap, net::SocketAddr};
 
-use blockchain::config::ConsensusConfig;
-use blockchain::config::ValidatorKeys;
+use blockchain::config::{ConsensusConfig, ValidatorKeys};
 use crypto::{PublicKey, SecretKey};
 
 /// Abstract configuration.