+++ exonum/src/helpers/types.rs
@@ -14,9 +14,7 @@
 
 //! Common widely used type definitions.
 
-use std::fmt;
-use std::num::ParseIntError;
-use std::str::FromStr;
+use std::{fmt, num::ParseIntError, str::FromStr};
 
 use serde::{Deserialize, Deserializer, Serialize, Serializer};